password reset 처리 구체화

+++ api/user/serializers/update.py
@@ -159,8 +159,8 @@ class PasswordResetVerifierCreateSerializer(serializers.ModelSerializer):
         return attrs
 
     def send_simple_message(self, attrs):
-        # body = f'https://dev-change.net/api/v1/user/password-reset/%s/%s' % (attrs['code'], attrs['token'])
-        body = f'http://localhost:8000/api/v1/user/password-reset/%s/%s' % (attrs['code'], attrs['token'])
+        body = f'https://dev-change.net/api/v1/user/password-reset/%s/%s' % (attrs['code'], attrs['token'])
+        # body = f'http://localhost:8000/api/v1/user/password-reset/%s/%s' % (attrs['code'], attrs['token'])
         EmailLog.objects.create(to=attrs['email'], body=body, title="어라운드어스 비밀번호 재설정 링크")
 
 
@@ -169,52 +169,58 @@ class PasswordResetSerializer(serializers.Serializer):
 
 
 class PasswordResetConfirmSerializer(serializers.Serializer):
-    code = serializers.CharField(write_only=True)
-    email_token = serializers.CharField(write_only=True)
-    password = serializers.CharField(write_only=True)
-    password_confirm = serializers.CharField(write_only=True)
+    code = serializers.CharField(required=True)
+    email_token = serializers.CharField(required=True)
+    password = serializers.CharField(write_only=True, required=True)
+    password_confirm = serializers.CharField(write_only=True, required=True)
+
+    error = serializers.ListField(read_only=True)
 
     def validate(self, attrs):
         code = attrs.get('code')
-        email_token = attrs.pop('email_token', None)
+        email_token = attrs.get('email_token')
         password = attrs.get('password')
-        password_confirm = attrs.pop('password_confirm', None)
+        password_confirm = attrs.get('password_confirm')
         # 이메일 토큰 검증
         try:
             self.email_verifier = EmailVerifier.objects.get(code=code, token=email_token)
         except EmailVerifier.DoesNotExist:
-            raise ValidationError('유효한 링크가 아닙니다.')
+            attrs.update({
+                'error': [0, '유효한 링크가 아닙니다.']
+            })
 
         # 이메일 검증
         if not User.objects.filter(email=self.email_verifier.email).exists():
-            raise ValidationError({'email': ['존재하지 않는 이메일입니다.']})
+            attrs.update({
+                'error': [0, '유효한 링크가 아닙니다.']
+            })
         self.email = self.email_verifier.email
 
         # password 검증
-        errors = {}
         if password != password_confirm:
-            errors['password'] = ['비밀번호가 일치하지 않습니다.']
-            errors['password_confirm'] = ['비밀번호가 일치하지 않습니다.']
+            attrs.update({
+                'error': [1, '비밀번호가 일치하지 않습니다.']
+            })
         else:
             try:
                 validate_password(password)
             except DjangoValidationError as error:
-                errors['password'] = list(error)
-                errors['password_confirm'] = list(error)
-
-        if errors:
-            raise ValidationError(errors)
+                print(error)
+                attrs.update({
+                    'error': [1, list(error)[0]]
+                })
 
         return attrs
 
     @transaction.atomic
     def create(self, validated_data):
-        password = validated_data.pop('password')
+        if validated_data.get('error') is None:
+            password = validated_data.pop('password')
 
-        user = User.objects.get(email=self.email)
-        user.set_password(password)
-        user.save()
+            user = User.objects.get(email=self.email)
+            user.set_password(password)
+            user.save()
 
-        self.email_verifier.delete()
+            self.email_verifier.delete()
 
-        return user
+        return validated_data
